chore: automate releases
This also fixes our current dependency on GitHub archives which are not guaranteed to be stable,
https://github.com/bazelbuild/bazel-central-registry/pull/332/files#diff-7288b02983f19cbe12177d5f193a6bfc533e352fc44fbe12cdb93e1c8c28dcd1R4
which requires a special 'skip-url-stability-check' label on the BCR PR

Instead, we run 'git archive' to create the release artifact. We can now prune the tests/ folder from that artifact, which lets us remove an exclusion from the presubmit as well.